Colors for styling
I've removed the dark green color that was in my logo. Michael advised to use the light green as it looks better. 
Gray colors are optional if you want depending on what we want to use them so we may discuss that so we're all on the same page in regards to that.